% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcdd.R
\name{load_dictionary_file}
\alias{load_dictionary_file}
\title{Load the dictionary}
\usage{
load_dictionary_file(path, ...)
}
\arguments{
\item{path}{Path to the dictionary}

\item{...}{additional parameters passed to rio::import and its underlying functions.
For example the sheet name for an excel file}
}
\value{
a tibble containing the dictionary
}
\description{
Load the dictionary
}
\keyword{internal}
